[WIP] Enable textures sample on Metal 4
This commit enables Metal 4 on the textures sample. For simplicity, it
still uses a MTL::CommandBuffer and MTL::BlitCommandEncoder to create
the KTX textures. Its possible that this could be re-written using
MTL4::ComputerCommandEncoder but its not clear if there are any benefits
as it would require custom synchronization to ensure the committed work
finishes before continuing sample startup.