Add support for Metal 4 (#17)
* [WIP] Metal 4 support

* Ensure buffers are included in residency set

This commit makes sure to add the buffers to the residency set
allocations. This fixes an issue seen when attempting to capture a Metal
GPU capture where there was no geometry information as the buffers
were missing from the residency set.

* More fixes and added comments about resizing frame buffer

* [WIP] Enable textures sample on Metal 4

This commit enables Metal 4 on the textures sample. For simplicity, it
still uses a MTL::CommandBuffer and MTL::BlitCommandEncoder to create
the KTX textures. Its possible that this could be re-written using
MTL4::ComputerCommandEncoder but its not clear if there are any benefits
as it would require custom synchronization to ensure the committed work
finishes before continuing sample startup.

* Disabled imgui for now until Metal 4 support

This commit cleans up the textures sample to render different textures using instance ID rather than the selection from ImGUI UI, as the ImGui support has been disabled until Metal 4 support is added to the backend.

* Fixed some leaks